+++ src/components/Icon.js
@@ -1,8 +1,10 @@
 function Icon(props) {
-    return <span className="icon" title={props.title}>
-        <span className={`material-icons ${props.art}`} aria-hidden>{props.art}</span>
-        <span>{props.title}</span>
+  return (
+    <span className="icon" title={props.title}>
+      <span className={`material-icons ${props.art}`} aria-hidden>{props.art}</span>
+      <span>{props.title}</span>
     </span>
+  );
 }
 
-export default Icon
+export default Icon;