+++ src/pages/Error.js
@@ -1,10 +1,15 @@
-import React from "react"
+import React from 'react';
 
-function Error() {
-    return <main>
-        <h1>Error</h1>
-        <div className="help">The page you are looking for does not exist!</div>
+function Error(props) {
+  return (
+    <main>
+      <h1>Error</h1>
+      <div className="help">{props.msg
+        ? props.msg
+        : "The page you are looking for does not exist!"
+      }</div>
     </main>
+  );
 }
 
-export default Error
+export default Error;