+++ src/pages/Home.js
@@ -1,22 +1,12 @@
-import React from "react"
-import db from "../static/data.ts"
-
-import Card from "../components/Card"
+import React from 'react';
+import List from '../components/List';
 
 function Home() {
-
-    function renderList() {
-        const db_arr = Object.entries(db), list = []
-
-        db_arr.forEach(([id, rest]) => list.push(<Card key={id} id={id} {...rest} />))
-
-        return <div className="list-container">{list}</div>
-    }
-
-
-    return <main>
-        {renderList()}
+  return (
+    <main>
+      <List />
     </main>
+  );
 }
 
-export default Home
+export default Home;